Split myresults into results module and example driver script.
